fix: reliability — MDCT double precision, Floor0 bark scale precision, Codebook null safety

+++ NVorbis/Floor0.cs
@@ -66,21 +66,21 @@ namespace NVorbis
 
         int[] SynthesizeBarkCurve(int n)
         {
-            var scale = _bark_map_size / toBARK(_rate / 2);
+            var scale = _bark_map_size / toBARK(_rate / 2.0);
 
             var map = new int[n + 1];
 
             for (int i = 0; i < n - 1; i++)
             {
-                map[i] = Math.Min(_bark_map_size - 1, (int)Math.Floor(toBARK((_rate / 2f) / n * i) * scale));
+                map[i] = Math.Min(_bark_map_size - 1, (int)Math.Floor(toBARK((_rate / 2.0) / n * i) * scale));
             }
             map[n] = -1;
             return map;
         }
 
-        static float toBARK(double lsp)
+        static double toBARK(double lsp)
         {
-            return (float)(13.1 * Math.Atan(0.00074 * lsp) + 2.24 * Math.Atan(0.0000000185 * lsp * lsp) + .0001 * lsp);
+            return 13.1 * Math.Atan(0.00074 * lsp) + 2.24 * Math.Atan(0.0000000185 * lsp * lsp) + .0001 * lsp;
         }
 
         float[] SynthesizeWDelMap(int n)
